qlv: move log model to qemu.qlog
